(webdriverio): add throttleCPU and throttleNetwork commands
