+++ packages/webdriverio/src/commands/browser.ts
@@ -29,7 +29,8 @@ export * from './browser/setCookies.js'
 export * from './browser/setTimeout.js'
 export * from './browser/setWindowSize.js'
 export * from './browser/switchWindow.js'
-export * from './browser/throttle.js'
+export * from './browser/throttleNetwork.js'
+export * from './browser/throttleCPU.js'
 export * from './browser/touchAction.js'
 export * from './browser/uploadFile.js'
 export * from './browser/url.js'
